requirements.txt
absl-py==2.0.0
astunparse==1.6.3
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
cryptography==41.0.4
et-xmlfile==1.1.0
flatbuffers==23.5.26
gast==0.5.4
google-auth==2.23.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.59.0
h5py==3.10.0
idna==3.4
keras==2.14.0
libclang==16.0.6
Markdown==3.5
MarkupSafe==2.1.3
ml-dtypes==0.2.0
numpy==1.26.1
oauthlib==3.2.2
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.1
Pillow==10.1.0
protobuf==4.24.4
psycopg2==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
python-dateutil==2.8.2
pytz==2023.3.post1
reportlab==4.0.6
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow-estimator==2.14.0
tensorflow-intel==2.14.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
Werkzeug==3.0.1
wrapt==1.14.1
XlsxWriter==3.1.9
